pyproject.toml
[tool.poetry]
name = "configclasses"
version = "0.1.0"
description = "Like dataclasses but for config."
authors = ["Pablo Cabezas <pabcabsal@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
tomlkit = {version = "^0.5.8", extras = ["toml"]}
python-dotenv = {version = "^0.10.3", extras = ["dotenv"]}
pyyaml = {version = "^5.1", extras = ["yaml"]}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
